requirements.txt
streamlit
supabase
groq
pandas
plotly
python-dotenv
fpdf2
